fix: address Copilot review comments and CI test failures
- useDebouncedCallback: use TArgs extends unknown[] constraint (no any), assign fnRef during render, add unmount cleanup, fix JSDoc
- highlightMatch: verify prefix/exact matches before highlighting, tighten matchType to union type
- formatDate: make medium style distinct ("5 January 2025" vs short's "5 Jan 2025")
- RulesBrowserPage.test: add missing formatDate/useDebouncedCallback to @pops/ui mock
- TransactionsResultComponent.test: update expected income amount to include thousands separator ($3,500.00)
- PhotoUpload.test: update expected formatBytes output to include space and decimal (5.0 MB, 800.0 KB)

+++ packages/ui/src/lib/format.ts
@@ -36,9 +36,9 @@ export type DateStyle = 'short' | 'medium' | 'long' | 'datetime';
 /**
  * Format a date string or ISO timestamp.
  *
- * - `short`    — "5 Jan 2025"   (en-AU, day/month/year)
- * - `medium`   — "Jan 5, 2025"  (en-AU, month/day/year)
- * - `long`     — "Sunday, 5 January" (en-AU, weekday/month/day)
+ * - `short`    — "5 Jan 2025"      (en-AU, abbreviated month)
+ * - `medium`   — "5 January 2025"  (en-AU, full month name)
+ * - `long`     — "Sunday, 5 January" (en-AU, weekday + day + month)
  * - `datetime` — locale date+time string
  */
 export function formatDate(dateStr: string, style: DateStyle = 'short'): string {
@@ -53,7 +53,7 @@ export function formatDate(dateStr: string, style: DateStyle = 'short'): string 
     case 'medium':
       return date.toLocaleDateString('en-AU', {
         year: 'numeric',
-        month: 'short',
+        month: 'long',
         day: 'numeric',
       });
     case 'long':
